hook usePopularPeopleData renamed to usePopularPeople
